+++ relay-server/src/services/buffer/mod.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 
 use relay_base_schema::project::ProjectKey;
 use relay_config::Config;
-use tokio::sync::{Mutex, MutexGuard};
+use tokio::sync::MutexGuard;
 
 use crate::envelope::Envelope;
 
@@ -12,7 +12,20 @@ mod envelopestack;
 
 /// Wrapper for the EnvelopeBuffer implementation.
 #[derive(Debug)]
-pub struct EnvelopeBuffer(Arc<Mutex<dyn envelopebuffer::EnvelopeBuffer>>);
+pub struct EnvelopeBuffer(
+    /// TODO: Reconsider synchronization mechanism.
+    /// We can either
+    /// - keep the interface sync and use a std Mutex. In this case, we create a queue of threads.
+    /// - use an async interface with a tokio mutex. In this case, we create a queue of futures.
+    /// - use message passing (service or channel). In this case, we create a queue of messages.
+    ///
+    /// From the tokio docs:
+    ///
+    /// >  The primary use case for the async mutex is to provide shared mutable access to IO resources such as a database connection.
+    /// > [...] when you do want shared access to an IO resource, it is often better to spawn a task to manage the IO resource,
+    /// > and to use message passing to communicate with that task.
+    Arc<tokio::sync::Mutex<dyn envelopebuffer::EnvelopeBuffer>>,
+);
 
 impl EnvelopeBuffer {
     pub fn from_config(config: &Config) -> Self {
